requirements.txt
python=3.8
streamlit==0.89.0
pandas==2.0.3
snowflake-connector-python==3.7.0
streamlit_option_menu==0.3.12